moved testCreateChildAccount from lib/utils to e2e. committing with no changes made to give diff record of test changes
