+++ e2e/api/api_test.go
@@ -1,6 +1,7 @@
 package api_test
 
 import (
+	"encoding/json"
 	"io/ioutil"
 	"math/rand"
 	"os"
@@ -9,7 +10,9 @@ import (
 	"time"
 
 	"github.com/status-im/status-go/e2e"
+	"github.com/status-im/status-go/geth/account"
 	"github.com/status-im/status-go/geth/api"
+	"github.com/status-im/status-go/geth/common"
 	"github.com/status-im/status-go/geth/log"
 	"github.com/status-im/status-go/geth/params"
 	. "github.com/status-im/status-go/testing"
@@ -185,4 +188,140 @@ func (s *APITestSuite) TestLogoutRemovesCells() {
 
 	_, err = s.api.JailManager().Cell(testChatID)
 	require.Error(err, "Expected that cells was removed")
+}
+
+func testCreateChildAccount(t *testing.T) bool { //nolint: gocyclo
+	// to make sure that we start with empty account (which might get populated during previous tests)
+	if err := statusAPI.Logout(); err != nil {
+		t.Fatal(err)
+	}
+
+	keyStore, err := statusAPI.NodeManager().AccountKeyStore()
+	if err != nil {
+		t.Error(err)
+		return false
+	}
+
+	// create an account
+	createAccountResponse := common.AccountInfo{}
+	rawResponse := CreateAccount(C.CString(TestConfig.Account1.Password))
+
+	if err = json.Unmarshal([]byte(C.GoString(rawResponse)), &createAccountResponse); err != nil {
+		t.Errorf("cannot decode CreateAccount response (%s): %v", C.GoString(rawResponse), err)
+		return false
+	}
+
+	if createAccountResponse.Error != "" {
+		t.Errorf("could not create account: %s", err)
+		return false
+	}
+	address, pubKey, mnemonic := createAccountResponse.Address, createAccountResponse.PubKey, createAccountResponse.Mnemonic
+	t.Logf("Account created: {address: %s, key: %s, mnemonic:%s}", address, pubKey, mnemonic)
+
+	acct, err := common.ParseAccountString(address)
+	if err != nil {
+		t.Errorf("can not get account from address: %v", err)
+		return false
+	}
+
+	// obtain decrypted key, and make sure that extended key (which will be used as root for sub-accounts) is present
+	_, key, err := keyStore.AccountDecryptedKey(acct, TestConfig.Account1.Password)
+	if err != nil {
+		t.Errorf("can not obtain decrypted account key: %v", err)
+		return false
+	}
+
+	if key.ExtendedKey == nil {
+		t.Error("CKD#2 has not been generated for new account")
+		return false
+	}
+
+	// try creating sub-account, w/o selecting main account i.e. w/o login to main account
+	createSubAccountResponse := common.AccountInfo{}
+	rawResponse = CreateChildAccount(C.CString(""), C.CString(TestConfig.Account1.Password))
+
+	if err = json.Unmarshal([]byte(C.GoString(rawResponse)), &createSubAccountResponse); err != nil {
+		t.Errorf("cannot decode CreateChildAccount response (%s): %v", C.GoString(rawResponse), err)
+		return false
+	}
+
+	if createSubAccountResponse.Error != account.ErrNoAccountSelected.Error() {
+		t.Errorf("expected error is not returned (tried to create sub-account w/o login): %v", createSubAccountResponse.Error)
+		return false
+	}
+
+	err = statusAPI.SelectAccount(address, TestConfig.Account1.Password)
+	if err != nil {
+		t.Errorf("Test failed: could not select account: %v", err)
+		return false
+	}
+
+	// try to create sub-account with wrong password
+	createSubAccountResponse = common.AccountInfo{}
+	rawResponse = CreateChildAccount(C.CString(""), C.CString("wrong password"))
+
+	if err := json.Unmarshal([]byte(C.GoString(rawResponse)), &createSubAccountResponse); err != nil {
+		t.Errorf("cannot decode CreateChildAccount response (%s): %v", C.GoString(rawResponse), err)
+		return false
+	}
+
+	if createSubAccountResponse.Error != "cannot retrieve a valid key for a given account: could not decrypt key with given passphrase" {
+		t.Errorf("expected error is not returned (tried to create sub-account with wrong password): %v", createSubAccountResponse.Error)
+		return false
+	}
+
+	// create sub-account (from implicit parent)
+	createSubAccountResponse1 := common.AccountInfo{}
+	rawResponse = CreateChildAccount(C.CString(""), C.CString(TestConfig.Account1.Password))
+
+	if err := json.Unmarshal([]byte(C.GoString(rawResponse)), &createSubAccountResponse1); err != nil {
+		t.Errorf("cannot decode CreateChildAccount response (%s): %v", C.GoString(rawResponse), err)
+		return false
+	}
+
+	if createSubAccountResponse1.Error != "" {
+		t.Errorf("cannot create sub-account: %v", createSubAccountResponse1.Error)
+		return false
+	}
+
+	// make sure that sub-account index automatically progresses
+	createSubAccountResponse2 := common.AccountInfo{}
+	rawResponse = CreateChildAccount(C.CString(""), C.CString(TestConfig.Account1.Password))
+
+	if err := json.Unmarshal([]byte(C.GoString(rawResponse)), &createSubAccountResponse2); err != nil {
+		t.Errorf("cannot decode CreateChildAccount response (%s): %v", C.GoString(rawResponse), err)
+		return false
+	}
+
+	if createSubAccountResponse2.Error != "" {
+		t.Errorf("cannot create sub-account: %v", createSubAccountResponse2.Error)
+	}
+
+	if createSubAccountResponse1.Address == createSubAccountResponse2.Address || createSubAccountResponse1.PubKey == createSubAccountResponse2.PubKey {
+		t.Error("sub-account index auto-increament failed")
+		return false
+	}
+
+	// create sub-account (from explicit parent)
+	createSubAccountResponse3 := common.AccountInfo{}
+	rawResponse = CreateChildAccount(C.CString(createSubAccountResponse2.Address), C.CString(TestConfig.Account1.Password))
+
+	if err := json.Unmarshal([]byte(C.GoString(rawResponse)), &createSubAccountResponse3); err != nil {
+		t.Errorf("cannot decode CreateChildAccount response (%s): %v", C.GoString(rawResponse), err)
+		return false
+	}
+
+	if createSubAccountResponse3.Error != "" {
+		t.Errorf("cannot create sub-account: %v", createSubAccountResponse3.Error)
+	}
+
+	subAccount1, subAccount2, subAccount3 := createSubAccountResponse1.Address, createSubAccountResponse2.Address, createSubAccountResponse3.Address
+	subPubKey1, subPubKey2, subPubKey3 := createSubAccountResponse1.PubKey, createSubAccountResponse2.PubKey, createSubAccountResponse3.PubKey
+
+	if subAccount1 == subAccount3 || subPubKey1 == subPubKey3 || subAccount2 == subAccount3 || subPubKey2 == subPubKey3 {
+		t.Error("sub-account index auto-increament failed")
+		return false
+	}
+
+	return true
 }
